Updated releng/autorelease Project: ovsdb  ec57b0954b0fbfe6b4f3553c65ab2aa39d3b9490
Bug 4331 - External bridge used for gateway mac resolver needs to be dynamic (cont.)

This is an additional patchset for changes that shuffle list of dpids so that
different br-ex instances are used when mac resolver is not getting an answer.
The other part of this gerrit has been merged via:
https://git.opendaylight.org/gerrit/#/c/32284/

With this change the GatewayMacResolverService will become resilient to the cases
where nodes get deleted or lose br-ex connectivity after the resolver is kicked
off to periodically track a given ip address.

Also made some minor clean up in NeutronL3Adapter, so that it needs not to
duplicate work for tracking ip in the very first iteration. That duplication
is not necessary as the GatewayMacResolverService is already implemented
with thread executors for the periodic monitoring.

Lastly, this gerrit augments the GatewayMacResolver service to provide
an extra boolean param called refreshExternalNetworkBridgeDpidIfNeeded.
That can be used to control whether dpid used gets updated when mac is
not getting resolved.

Change-Id: I142735f875d1a79144c204bec631712de00d9746
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>