Updated releng/autorelease Project: ovsdb  b1cf5de78228b23a4c6aae1613df925acfa798c9
Bug 4927 - Rules added by MacResolverServices in br-ex are not removed

This bug can be reproduced by using an external router address
that is not owned by any system in the external neutron subnet.

Made the flow remove call idem-potent and in a common function.

Added missing flow remove calls in the following 2 missing places:
  * Entry removal (non periodic and periodic)
  * Time period restart

Change-Id: Idf481d3a13c1728bf400f185de296cbada7d920b
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>