Updated releng/autorelease Project: ovsdb  c20313b4a4fb6340d4291ce71ee37eea95bb7f7f
Enable PowerMock instrumentation with JaCoCo

By default PowerMock is incompatible with JaCoCo, even with the older
version of JaCoCo defined by odlparent: the tests run but their
coverage isn't measured. This patch forces offline instrumentation
which allows both PowerMock and JaCoCo to operate (more or less)
properly. There are a number of "class already instrumented" errors
but they don't affect the test results.

This patch also upgrades PowerMock to 1.6.4 (initially in the hope
that it would work with JaCoCo as-is); this requires an upgrade to
Mockito too. I'll drive these changes properly in odlparent and remove
the local overrides when appropriate!

Change-Id: I2ab704f7a28f325ed3bb5a75a6f34ffb6c3b36b3
Signed-off-by: Stephen Kitt <skitt@redhat.com>