Updated releng/autorelease Project: yangtools  fce70654f7d387a8454c8441459ffff03ad0f1a2
BUG-4295: fix merge callsite

This is a fixup of the previous patch. The problem was expanding a merge
into a new node, which used a direct value propagation with a merge.
That is not correct, as the merge should be pushed via
ModificationApplyOperation, which in case of unkeyed lists will turn it
into a write.

Change-Id: I93c8be80e4467b5d3a1e20f26f5576bd701b1375
Signed-off-by: Robert Varga <rovarga@cisco.com>
(cherry picked from commit dfe3c8193b6fd06090de7221e51e5857b75517b6)