Updated releng/autorelease Project: openflowplugin  636fd4b5cdf0751c684b0c6996523a9fedb5c63a
Send node details notification up to the application,
once switch role is configured. Currently application were getting notified when switch Description message
received by the controller, and application assume that switch is configured and ready to use.
This patch, send the desc request when switch role is configured and it's ready to use

Change-Id: I8a02c209c4796b7aa01161257d523139fdc9c881
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>