Updated releng/autorelease Project: neutron  59f3919dd01797fb056fa5dcdebda2076a936954
BUG-5137: extraroutes in router not working

Neutron API defines extraroutes as list of destination and nexthop but in
neutron-l3.yang they're defined as list of strings. So, when these routes
are present in a router object sent by openstack we are unable to store
them in MDSAL. Fix is to change the type of routes in l3-attributes to
match with what Neutron API sends us.

Change-Id: I83519edcbdf6f4fcae34ae9650cf8ede40256296
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>