Updated releng/autorelease Project: controller  967558bda496952706e78581b7a185e11e66806e
Bug 5175: RPC called on mount point not being correctly handled when multiple revision-date coexist

The ControllerContext#toQname was just looking for module using the module name.
In scenarios where you have multiple revision for the same module, the method
was returning the latest schema.

This fix is to retreive the schema based on the revision date

Change-Id: Ief10dab8925bc942f446895fe7078ebcb1105a3a
Signed-off-by: adetalhouet <adetalhouet@inocybe.com>