Updated releng/autorelease Project: aaa  83b3c36d60b56eb5c9f2a105405418105dfce146
Ensure H2 resources are closed

This patch uses try-with-resources with all database resources
(connections, statements and result sets) to ensure they're closed
correctly in all cases. It drops the re-used dbConnection since that
seems fragile (two threads accessing the store simultaneously might
get the same connection, and one of the threads will close it before
the other has finished), except for tests.

The initial table check/creation is synchronized to avoid
time-of-check to time-of-use races.

Common code is extracted to an AbstractStore.

Exceptions are logged and re-thrown as StoreExceptions with exception
chaining.

Change-Id: Ia63493fcb1361e53a5f3400ee5e2fdf09bccb574
Signed-off-by: Stephen Kitt <skitt@redhat.com>