Updated releng/autorelease Project: lispflowmapping  a5fc5fb68dc91eb1dec1339c53271535e3c37a2d
Bug 5239: Fix IID Source/Dest lookup

The issue is that the source EID is passed to the map-cache lookup as an
IP, not as an IP-prefix, and therefore it is not masked accordingly when
performing the longest prefix match. This patch updates the
MaskUtil.normalize() code to convert IPs to an IP-prefixes.

Change-Id: Iafd7bef2ef0911e6cbc61b8315208eb639e41766
Signed-off-by: Florin Coras <fcoras@cisco.com>