Updated releng/autorelease Project: tsdr  b89c975ea0f92f11e3e70c8aaefdf9e36f2f4e95
Fix of Bug 5244

This patch set fixed two issues described in Bug 5244:
1.Prefix issue in hbase data store.
Currently when the user queries for [NID=openflow:1], the query
result will contain records that matches NID=openflow:111 too.
This fix solved this issue to return the exact match instead of
prefix match.

When user uses tsdrkey to query data from tsdr data store, each
section in tsdrkey is optional except the data category [DC=...].
However, for the content inside each section, the match has to be
the exact match instead of fuzzy match. This also applies to the
[RK=..] recordkey section. The RecordKey has to be exact match.
This means, if RK contains NodeID, FlowTableID, and FlowID, the
user has to supply the full key in RK section. For example, [RK=
Node:openflow:1,Table:0,Flow:6] is a legal query. [RK=Node:openflow:1]
would not return any results since the recordkey is partial.

2. When the size of HBase data store is big enough, the filtering
does not work.

The root cause was the PageFilter that was set on the scanner.
When the PageFilter was set with 1000 records, only 1000 records
will be returned. Since the logic was applying the filter on the
returned records, when the hbase table size is larger than 1000,
if the target row is beyond 1000, the matched record won't be
returned.

The fix removed the PageFilter on the scanner. Since we've already
applied the read caching on the scanner, the performance has
been optimized. The PageFilter would not help with the performance.
Instead it caused matched data not being returned to the client.

Change-Id: I927a0e5d4b1b3ee0bcc6afe4c33e30692462e34a
Signed-off-by: YuLing Chen <yulingchen54@gmail.com>