Updated releng/autorelease Project: openflowplugin  d342b69bcd7bb2a7ad10316821025f774459b76f
Bug 5050: Correctly deregister routed RPC.

OfEntityManager of the OpenFlow plugin (the Helium version) forgot to
update the flag which represented if it had registered its routed RPC,
when it deregistered the routed RPC. As the result, when the OpenFlow
plugin tried to register and deregister the routed RPC more than once,
it wrongly skipped to reregister the routed RPC from the second time of
the registeration, since it thought the RPC had been already registered.
When the plugin failed to get the response to a Role Request message,
the plugin faced such a situation.

Change-Id: I3f5bf59e254e58f20649ec205656a03debdd4eb5
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>
(cherry picked from commit 8566438a6f43ee39c22d70b0c13327d9da66f397)