Updated releng/autorelease Project: vpnservice  948cc8428f13c0929c190ed6d0abac2eab3f6ae3
Use {{VERSION}} in features and drop unneeded properties

{{VERSION}} can replace many properties in features.xml (the versions
are then taken from the POM).

commons-net is provided in odlparent's dependency management, the
commons.net.version property can be safely dropped.

With the upgrade of Mockito in odlparent, PowerMock needs to be
upgraded to 1.6.4; we simply rely on odlparent to provide the correct
version.

Change-Id: Iddd2fcc429e95d7ffd81f314fe183d138dedeefd
Signed-off-by: Stephen Kitt <skitt@redhat.com>