Updated releng/autorelease Project: nic  0624f87a4a837156474e698c0829bd3493063f38
Bug 5389: Put intents to OPERATIONAL data store.

When VTN Renderer successfully renders an intent, it needs to update the
OPERATIONAL data store.  However, the VTN Renderer tried to update the
CONFIGURATION data store. The patch fixed this issue, and also improved
the way to update the data store.

Other changes:

  * Fixed wrong comments.
  * Deleted the unused private method and variables.

Change-Id: I541f64396301b5a160fa84380bfa696929124388
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>