Updated releng/autorelease Project: sfc  df51faf84ca47049cbd1791edc3891383e89c81b
Clean up dependency properties

Migrate to features-parent for features-sfc and use that to drop a
number of dependency properties (pulling in the dependency management
from odlparent).

Remove obsolete properties.

Note that this upgrades PowerMock to 1.6.4 (odlparent now provides
dependency management for Mockito and PowerMock since they need to be
upgaded in sync).

Change-Id: Ib0fc1a63f16867082bb3eb12ee51cc89ecc7a3bc
Signed-off-by: Stephen Kitt <skitt@redhat.com>