Updated releng/autorelease Project: ovsdb  c1d115998e5a7561c0b3c047f8ed003e2d036ddc
Do not duplicate MacAddress objects

MacAddress is an immutable object, hence it does not make sense to
duplicate it. Same is true for IpvXPrefix.

Change-Id: Ic3e4ec8cc66e90ba13e93d47630d104d2188f738
Signed-off-by: Robert Varga <rovarga@cisco.com>

Use IetfInetUtil for IPv4 address creation

Optimized version for instantiating a IpAddress from InetAddress. There
is also a IPv6 version of this, but that produces differently-formatted
addresses and break strict backwards compatibility.

Change-Id: I40d2204b5709e5fa6fe2b98af86a893a07b36862
Signed-off-by: Robert Varga <rovarga@cisco.com>

Use InetAddresses to instantiate IPv4 addresses

Unlike .getByName(), these do not trigger resolver. The IPv6 version
unmasks mapped addresses, hence it is not used here.

Change-Id: I5164f28cab29a4fec624d906fa198ffa21dd0e87
Signed-off-by: Robert Varga <rovarga@cisco.com>