Updated releng/autorelease Project: lispflowmapping  988a18f958a41fed647f260a0aa56a09e9c48618
Create typedefs for xTR-ID and Site-ID

There were a few places where we had site-id and xtr-id leaves, an they
used binary types, not a common type. This patch creates a typedef for
the two data types and uses it consistently throughout the code, for
better type safety.

The patch is in preparation for xTR-ID data store support for
southbound.

Change-Id: Ia45541b4f82fd2697e05686d995a219524e49e99
Signed-off-by: Lorand Jakab <lojakab@cisco.com>