Updated releng/autorelease Project: ovsdb  a4e7a0d8eadeb1e5db154be6536f046f8b328f6f
Net-virt is not checking if the master of the net-virt-application is
elected at the clustering level or not. Because of this it can end up in a situation
where candidates are registered but owner is not elected and all the instances
gets false when they check for the owner. This patch fixes the issue, where
it waits till owner is selected

Change-Id: I2c720aeac44e4252091952383fe3bb4ed9bed97b
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>