Updated releng/autorelease Project: lispflowmapping  3b92f87adc6d0e1bae582398bdaf9e5816040b0b
Integration tests: turn on debugging in Karaf log

This patchs turns on karaf debug output for integration tests for all
lispflowmapping classes, not just the MappingServiceIntegrationTest
class itself. This is extremely valuable when debugging integration
tests issues, both on a local build, or in Jenkins. The karaf log for
the integration tests can be found in the
integrationtest/target/surefire-reports/org.opendaylight.lispflowmapping.integrationtest.MappingServiceIntegrationTest-output.txt
file, and with this patch it will have full debug output. If trace level
output is needed, the next line can be changed too, but that's overkill
for most situations, and should be used at a developer's discretion, but
not by default in the build.

Change-Id: Ice723eac5509706e8b633bab5aa02d83758855fd
Signed-off-by: Lorand Jakab <lojakab@cisco.com>
(cherry picked from commit 045459dfc36478405fe621ca0e31573ecf79cd78)