Updated releng/autorelease Project: bgpcep  02559d3d5f4be9e716a30580186740f4fded6c8e
BUG-4827: Modify ipv4-routes list model, to include path-id as part of the key

Add path Id as a part of the key for ipv4-routes. This will force us to use
path-Id when writing on Loc-Rib even when add-path selection mode are not
used. Therefore we need to adapt the code on base selection mode
to take in consideration wheter model can support Add-path or not
and proced accordingly.

Change-Id: I94bc6523ff788c236008df892569d3ea04cadb65
Signed-off-by: Claudio D. Gasparini <cgaspari@cisco.com>